refactor(acp): extract ext-method handlers to ext-methods/mcp.ts
Address vmizg's review comments on PR #946:

- Extract validateServerEntry() and handleProbeMcpServer() from server.ts
  into src/modes/acp/ext-methods/mcp.ts. server.ts was ~1700 lines; the
  ext-method logic is self-contained and belongs in its own module.

- Convert extMethod() dispatch from if/cascade to switch/case for cleaner
  routing and extensibility.

- Remove unnecessary { ...result } shallow copy — the ProbeResult object
  crosses the ACP wire immediately and is not mutated, so the spread was
  defensive overhead with no benefit.
